Fixed Apple1IO so no characters are lost when toggling terminal speed, optimized Apple1Terminal, implemented Apple1Terminal splash screen, implemented clear screen on CTRL+L in Apple1Terminal, fixed Monitor bezel handling and canvas cleaning on power off, implemented automatic software update, optimized MOS6502 and W65C02S special condition handling

+++ libemulation/Implementation/Generic/RAM.cpp
@@ -87,11 +87,11 @@ bool RAM::init()
     if (!powerOnPattern.size())
         powerOnPattern.resize(1);
     else
-        powerOnPattern.resize(getNextPowerOf2((int) powerOnPattern.size()));
+        powerOnPattern.resize((size_t) getNextPowerOf2((int) powerOnPattern.size()));
     
     size = getNextPowerOf2(size);
     size_t oldSize = data.size();
-    data.resize(size);
+    data.resize((size_t) size);
     if (oldSize == 0)
         initMemory();
     datap = (OEUInt8 *) &data.front();
